split up server from server socket
