+++ transport/src/commonMain/kotlin/mqtt/transport/nio/socket/readStats.kt
@@ -1,0 +1,3 @@
+package mqtt.transport.nio.socket
+
+expect fun readStats(port: UShort, contains: String): List<String>
