+++ transport/src/commonTest/kotlin/mqtt/transport/SocketTests.kt
@@ -77,7 +77,7 @@ class SocketTests {
         val mutex = Mutex(true)
         var serverClientSocket: ClientSocket? = null
         serverSocket.bind()
-        val server = server(serverSocket)
+        val server = SocketServer(serverSocket)
         val launched = ServerLaunched(firstReceiveLock, server, serverSocket, serverClientSocket, mutex)
         scope.launch {
             server.listen().collect {
